Automatically determine cipher mode

+++ tricks/jwplayer.sh
@@ -17,17 +17,36 @@ try_jwplayer() {
 
 		# extract key
 		key=$(swfdump -D "$tmp" 2>/dev/null | grep -A4 'decryptor::h1' | grep String | sed 's/.*= //' | paste -sd' ' | tr -d ' ')
+		echo -n "$key" > ~/.cache/streamsh-jwfile.key
+
+		# determine encryption mode
+		mode=$(swfdump -D "$tmp" 2>/dev/null \
+			| grep -C1 "pushstring" \
+			| grep -C3 "setlocal r4" \
+			| grep -C2 "setlocal r5" \
+			| grep pushstring \
+			| sed 's/.*pushstring "\(.*\)"/\1/' \
+			| head -n1
+		)
+		if [[ -z "$mode" ]]; then
+			echo "Could not determine encryption mode :'(" > /dev/stderr
+			return
+		fi
+		cipher=$(echo "$mode" | sed 's/simple-//' | awk -F- '{print $1}')
+		chain=$(echo "$mode" | awk -F- '{print $NF}')
+		bits=$(echo -n "$key" | wc -c)
+		bits=$(echo "$bits*4" | bc -l)
 		rm "$tmp"
-		echo "$key" > ~/.cache/streamsh-jwfile.key
 
 		# get binary data
 		(echo -n "$url" | xxd -r -p; echo "") > ~/.cache/streamsh-jwfile.aes
 
 		# decrypt
+		echo "Decrypting with $cipher-$bits-$chain and key '$key'" > /dev/stderr
 		dec=$(
 			(echo -n "$url" | xxd -r -p; echo "") \
 			| base64 --wrap=0 \
-			| openssl enc -d -A -base64 -aes-128-ecb -K "$key" \
+			| openssl enc -d -A -base64 -${cipher}-${bits}-${chain} -K "$key" \
 			2>/dev/null
 		)
 		m=$(echo "$dec" | grep :)
